+++ src/utils/scales/scale_continuous.ts
@@ -103,8 +103,18 @@ interface ScaleOptions {
    * @default 10
    */
   ticks: number;
+  /** true if the scale was adjusted to fit one single value histogram */
+  isSingleValueHistogram: boolean;
 }
-
+const defaultScaleOptions: ScaleOptions = {
+  bandwidth: 0,
+  minInterval: 0,
+  timeZone: 'utc',
+  totalBarsInCluster: 1,
+  barsPadding: 0,
+  ticks: 10,
+  isSingleValueHistogram: false,
+};
 export class ScaleContinuous implements Scale {
   readonly bandwidth: number;
   readonly totalBarsInCluster: number;
@@ -118,22 +128,21 @@ export class ScaleContinuous implements Scale {
   readonly tickValues: number[];
   readonly timeZone: string;
   readonly barsPadding: number;
+  readonly isSingleValueHistogram: boolean;
   private readonly d3Scale: any;
 
   constructor(scaleData: ScaleData, options?: Partial<ScaleOptions>) {
     const { type, domain, range } = scaleData;
-    const scaleOptions: ScaleOptions = mergePartial(
-      {
-        bandwidth: 0,
-        minInterval: 0,
-        timeZone: 'utc',
-        totalBarsInCluster: 1,
-        barsPadding: 0,
-        ticks: 10,
-      },
-      options,
-    );
-    const { bandwidth, minInterval, timeZone, totalBarsInCluster, barsPadding, ticks } = scaleOptions;
+    const {
+      bandwidth,
+      minInterval,
+      timeZone,
+      totalBarsInCluster,
+      barsPadding,
+      ticks,
+      isSingleValueHistogram,
+    } = mergePartial(defaultScaleOptions, options);
+
     this.d3Scale = SCALES[type]();
     if (type === ScaleType.Log) {
       this.domain = limitLogScaleDomain(domain);
@@ -154,6 +163,7 @@ export class ScaleContinuous implements Scale {
     this.isInverted = this.domain[0] > this.domain[1];
     this.timeZone = timeZone;
     this.totalBarsInCluster = totalBarsInCluster;
+    this.isSingleValueHistogram = isSingleValueHistogram;
     if (type === ScaleType.Time) {
       const startDomain = DateTime.fromMillis(this.domain[0], { zone: this.timeZone });
       const endDomain = DateTime.fromMillis(this.domain[1], { zone: this.timeZone });
@@ -186,7 +196,12 @@ export class ScaleContinuous implements Scale {
   scale(value: any) {
     return this.d3Scale(value) + (this.bandwidthPadding / 2) * this.totalBarsInCluster;
   }
-
+  pureScale(value: any) {
+    if (this.bandwidth === 0) {
+      return this.d3Scale(value);
+    }
+    return this.d3Scale(value + this.minInterval / 2);
+  }
   ticks() {
     return this.tickValues;
   }
@@ -197,36 +212,49 @@ export class ScaleContinuous implements Scale {
     }
     return invertedValue;
   }
-  invertWithStep(value: number, data: number[]): any {
-    const invertedValue = this.invert(value - this.bandwidth / 2);
-    const leftIndex = bisectLeft(data, invertedValue);
+  invertWithStep(
+    value: number,
+    data: number[],
+  ): {
+    value: any;
+    withinBandwidth: boolean;
+  } {
+    const invertedValue = this.invert(value);
+    const bisectValue = this.bandwidth === 0 ? invertedValue + this.minInterval / 2 : invertedValue;
+    const leftIndex = bisectLeft(data, bisectValue);
+
     if (leftIndex === 0) {
-      // is equal or less than the first value
-      const prevValue1 = data[leftIndex];
-      if (data.length === 0) {
-        return prevValue1;
-      }
-      const nextValue1 = data[leftIndex + 1];
-      const nextDiff1 = Math.abs(nextValue1 - invertedValue);
-      const prevDiff1 = Math.abs(invertedValue - prevValue1);
-      if (nextDiff1 < prevDiff1) {
-        return nextValue1;
-      }
-      return prevValue1;
+      return {
+        value: data[0],
+        withinBandwidth: true,
+      };
     }
-    if (leftIndex === data.length) {
-      return data[leftIndex - 1];
+    const currentValue = data[leftIndex - 1];
+    // pure linear scale
+    if (this.minInterval === 0) {
+      const nextValue = data[leftIndex];
+      const nextDiff = Math.abs(nextValue - invertedValue);
+      const prevDiff = Math.abs(invertedValue - currentValue);
+      return {
+        value: nextDiff <= prevDiff ? nextValue : currentValue,
+        withinBandwidth: true,
+      };
     }
-    const nextValue = data[leftIndex];
-    const prevValue = data[leftIndex - 1];
-    const nextDiff = Math.abs(nextValue - invertedValue);
-    const prevDiff = Math.abs(invertedValue - prevValue);
-    if (nextDiff <= prevDiff) {
-      return nextValue;
+    if (invertedValue - currentValue <= this.minInterval) {
+      return {
+        value: currentValue,
+        withinBandwidth: true,
+      };
     }
-    return prevValue;
+    return {
+      value: currentValue + this.minInterval * Math.floor((invertedValue - currentValue) / this.minInterval),
+      withinBandwidth: false,
+    };
   }
   isSingleValue() {
+    if (this.isSingleValueHistogram) {
+      return true;
+    }
     if (this.domain.length < 2) {
       return true;
     }